Forth commit

+++ controllers/cards.js
@@ -1,4 +1,7 @@
+const mongoose = require('mongoose');
 const Card = require('../models/card');
+
+const { ObjectId } = mongoose.Types;
 
 module.exports.getCards = (req, res) => {
   Card.find({})
@@ -23,27 +26,57 @@ module.exports.createCard = (req, res) => {
 
 module.exports.deleteCard = (req, res) => {
   const { cardId } = req.params;
+  if (!ObjectId.isValid(cardId)) {
+    res.status(404).send({ message: `Нет карточки с id ${cardId}` });
+    return;
+  }
   Card.findOneAndDelete({ _id: cardId })
-    .then((card) => res.send({ data: card }))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        res.status(400).send({ message: `${err.message}` });
-      } else if (err.name.indexOf('Cast to ObjectId failed')) {
+    .then((card) => {
+      if (!card) {
         res.status(404).send({ message: `Нет карточки с id ${cardId}` });
-      } else {
-        res.status(500).send({ message: `${err.message}` });
+        return;
       }
+      res.send({ data: card });
+    })
+    .catch((err) => {
+      res.status(500).send({ message: `${err.message}` });
     });
 };
 
 module.exports.likeCard = (req, res) => {
-  const { cardId } = req.body;
+  const { cardId } = req.params;
+  if (!ObjectId.isValid(cardId)) {
+    res.status(404).send({ message: `Нет карточки с id ${cardId}` });
+    return;
+  }
   Card.findByIdAndUpdate(cardId, { $addToSet: { likes: req.user._id } }, { new: true })
-    .then((card) => res.send({ data: card }));
+    .then((card) => {
+      if (!card) {
+        res.status(404).send({ message: `Нет карточки с id ${cardId}` });
+        return;
+      }
+      res.send({ data: card });
+    })
+    .catch((err) => {
+      res.status(500).send({ message: `${err.message}` });
+    });
 };
 
 module.exports.dislikeCard = (req, res) => {
-  const { cardId } = req.body;
+  const { cardId } = req.params;
+  if (!ObjectId.isValid(cardId)) {
+    res.status(404).send({ message: `Нет карточки с id ${cardId}` });
+    return;
+  }
   Card.findByIdAndUpdate(cardId, { $pull: { likes: req.user._id } }, { new: true })
-    .then((card) => res.send({ data: card }));
+    .then((card) => {
+      if (!card) {
+        res.status(404).send({ message: `Нет карточки с id ${cardId}` });
+        return;
+      }
+      res.send({ data: card });
+    })
+    .catch((err) => {
+      res.status(500).send({ message: `${err.message}` });
+    });
 };
